// frontend/src/app/api/model/habit.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HabitFrequency } from './habitFrequency';


export interface Habit { 
    id: number;
    name: string;
    description?: string | null;
    frequency: HabitFrequency;
    createdAt: string;
    active: boolean;
    completedToday: boolean;
    /**
     * Today (DAILY) or current Monday-Sunday week (WEEKLY)
     */
    doneForCurrentPeriod: boolean;
    /**
     * Consecutive periods with a completion
     */
    currentStreak: number;
}
export namespace Habit {
}