// frontend/src/app/api/model/habitWrite.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { HabitFrequency } from './habitFrequency';


export interface HabitWrite { 
    name: string;
    description?: string | null;
    frequency: HabitFrequency;
}
export namespace HabitWrite {
}